+++ ariadne_codegen/client_generators/dependencies/async_base_client.py
@@ -88,7 +88,9 @@ class AsyncBaseClient:
         if variables:
             payload["variables"] = self._convert_dict_to_json_serializable(variables)
         content = json.dumps(payload, default=pydantic_encoder)
-        return await self.http_client.post(url=self.url, content=content)
+        return await self.http_client.post(
+            url=self.url, content=content, headers={"Content-Type": "application/json"}
+        )
 
     def get_data(self, response: httpx.Response) -> Dict[str, Any]:
         if not response.is_success: