+++ tests/client_generators/dependencies/test_async_base_client.py
@@ -185,6 +185,31 @@ async def test_execute_sends_payload_with_serialized_datetime_without_exception(
     assert content["variables"]["arg"] == arg_value.isoformat()
 
 
+@pytest.mark.asyncio
+async def test_execute_sends_request_with_correct_content_type(httpx_mock):
+    httpx_mock.add_response()
+    client = AsyncBaseClient(url="http://base_url")
+
+    await client.execute("query Abc { abc }", {})
+
+    request = httpx_mock.get_request()
+    assert request.headers["Content-Type"] == "application/json"
+
+
+@pytest.mark.asyncio
+async def test_execute_sends_request_with_extra_headers_and_correct_content_type(
+    httpx_mock,
+):
+    httpx_mock.add_response()
+    client = AsyncBaseClient(url="http://base_url", headers={"h_key": "h_value"})
+
+    await client.execute("query Abc { abc }", {})
+
+    request = httpx_mock.get_request()
+    assert request.headers["h_key"] == "h_value"
+    assert request.headers["Content-Type"] == "application/json"
+
+
 @pytest.mark.parametrize(
     "status_code, response_content",
     [